Specify a processor backend
